DockerMysql

+++ src/main/java/com/ba/dto/OrderItemDTO.java
@@ -12,8 +12,8 @@ import lombok.Setter;
 public class OrderItemDTO {
     private Long id;
     private ProductDTO product;
-    private Long piece;
-    private Long totalPrice;
+    private int piece;
+    private int totalPrice;
     private OrderDTO order;
     private String table;
 }
